requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
Pygments==2.19.2
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-telegram-bot==22.3
requests==2.32.3
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.4.0
yarl==1.20.1
